third_party/blink/web_tests/external/wpt/editing/run/insertparagraph_5001-6000-expected.txt

This is a testharness.js-based test.
Found 15 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "<ul><li><div>foo[]</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li><div>foo</div></li><li><div><br></div></li></ul>" but got "<ul><li><div>foo</div></li></ul>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "<ul><li><div>foo[]</ul>" queryCommandValue("defaultparagraphseparator") after
  assert_equals: Wrong result returned expected "p" but got "div"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "<div><p>foo[]</div>" queryCommandValue("defaultparagraphseparator") before
  assert_equals: Wrong result returned expected "p" but got "div"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "<div><div>foo[]</div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div><div>foo</div><div><br></div></div>" but got "<div><div>foo</div></div><div><br></div>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "<div><div>foo[]</div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div><div>foo</div><div><br></div></div>" but got "<div><div>foo</div></div><div><br></div>"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "<ol><li class=a id=x><p class=b id=y>foo[]</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li class=\\"a\\" id=\\"x\\"><p class=\\"b\\" id=\\"y\\">foo</p></li><li class=\\"a\\"><p class=\\"b\\"><br></p></li></ol>" but got "<ol><li class=\\"a\\" id=\\"x\\"><p class=\\"b\\" id=\\"y\\">foo</p></li><li class=\\"a\\" id=\\"x\\"><p class=\\"b\\" id=\\"y\\"><br></p></li></ol>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "<ol><li class=a id=x><p class=b id=y>foo[]</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li class=\\"a\\" id=\\"x\\"><p class=\\"b\\" id=\\"y\\">foo</p></li><li class=\\"a\\"><p class=\\"b\\"><br></p></li></ol>" but got "<ol><li class=\\"a\\" id=\\"x\\"><p class=\\"b\\" id=\\"y\\">foo</p></li><li class=\\"a\\" id=\\"x\\"><p class=\\"b\\" id=\\"y\\"><br></p></li></ol>"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "<div class=a id=x><div class=b id=y>foo[]</div></div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div class=\\"a\\" id=\\"x\\"><div class=\\"b\\" id=\\"y\\">foo</div><div class=\\"b\\"><br></div></div>" but got "<div class=\\"a\\" id=\\"x\\"><div class=\\"b\\" id=\\"y\\">foo</div><div class=\\"b\\" id=\\"y\\"><br></div></div>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "<div class=a id=x><div class=b id=y>foo[]</div></div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div class=\\"a\\" id=\\"x\\"><div class=\\"b\\" id=\\"y\\">foo</div><div class=\\"b\\"><br></div></div>" but got "<div class=\\"a\\" id=\\"x\\"><div class=\\"b\\" id=\\"y\\">foo</div><div class=\\"b\\" id=\\"y\\"><br></div></div>"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "<div class=a id=x><p class=b id=y>foo[]</div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div class=\\"a\\" id=\\"x\\"><p class=\\"b\\" id=\\"y\\">foo</p><p class=\\"b\\"><br></p></div>" but got "<div class=\\"a\\" id=\\"x\\"><p class=\\"b\\" id=\\"y\\">foo</p><p class=\\"b\\" id=\\"y\\"><br></p></div>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "<div class=a id=x><p class=b id=y>foo[]</div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div class=\\"a\\" id=\\"x\\"><p class=\\"b\\" id=\\"y\\">foo</p><p class=\\"b\\"><br></p></div>" but got "<div class=\\"a\\" id=\\"x\\"><p class=\\"b\\" id=\\"y\\">foo</p><p class=\\"b\\" id=\\"y\\"><br></p></div>"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "<ol><li class=a id=x><p class=b id=y>[]foo</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li class=\\"a\\" id=\\"x\\"><p class=\\"b\\" id=\\"y\\"><br></p></li><li class=\\"a\\"><p class=\\"b\\">foo</p></li></ol>" but got "<ol><li class=\\"a\\" id=\\"x\\"><p class=\\"b\\" id=\\"y\\"><br></p></li><li class=\\"a\\" id=\\"x\\"><p class=\\"b\\" id=\\"y\\">foo</p></li></ol>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "<ol><li class=a id=x><p class=b id=y>[]foo</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li class=\\"a\\" id=\\"x\\"><p class=\\"b\\" id=\\"y\\"><br></p></li><li class=\\"a\\"><p class=\\"b\\">foo</p></li></ol>" but got "<ol><li class=\\"a\\" id=\\"x\\"><p class=\\"b\\" id=\\"y\\"><br></p></li><li class=\\"a\\" id=\\"x\\"><p class=\\"b\\" id=\\"y\\">foo</p></li></ol>"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "<div class=a id=x><div class=b id=y>[]foo</div></div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div class=\\"a\\" id=\\"x\\"><div class=\\"b\\" id=\\"y\\"><br></div><div class=\\"b\\">foo</div></div>" but got "<div class=\\"a\\" id=\\"x\\"><div class=\\"b\\" id=\\"y\\"><br></div><div class=\\"b\\" id=\\"y\\">foo</div></div>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "<div class=a id=x><div class=b id=y>[]foo</div></div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div class=\\"a\\" id=\\"x\\"><div class=\\"b\\" id=\\"y\\"><br></div><div class=\\"b\\">foo</div></div>" but got "<div class=\\"a\\" id=\\"x\\"><div class=\\"b\\" id=\\"y\\"><br></div><div class=\\"b\\" id=\\"y\\">foo</div></div>"
Harness: the test ran to completion.
